Move Angle tests into src/si/angle.rs
