+++ src/si/angle.rs
@@ -140,3 +140,61 @@ mod tests {
         }
     }
 }
+
+#[cfg(all(test, feature = "std"))]
+mod trig {
+    storage_types! {
+        types: Float;
+
+        use num::{FromPrimitive, Zero};
+
+        use tests::*;
+
+        use si::angle as a;
+        use si::quantities::*;
+
+        use ::lib::f64::consts::PI;
+
+        #[test]
+        fn sanity() {
+
+            let zero: Angle<V> = Angle::zero();
+            let nzero: Angle<V> = -Angle::zero();
+            let pi: Angle<V> = Angle::new::<a::radian>(V::from_f64(PI).unwrap());
+            let half: Angle<V> = Angle::new::<a::radian>(V::from_f64(PI / 2.0).unwrap());
+
+
+            Test::assert_approx_eq(&zero.cos(), &1.0);
+            Test::assert_approx_eq(&nzero.cos(), &1.0);
+
+            Test::assert_approx_eq(&pi.cos(), &-1.0);
+            Test::assert_approx_eq(&half.cos(), &0.0);
+
+            Test::assert_approx_eq(&zero.sin(), &0.0);
+            Test::assert_approx_eq(&nzero.sin(), &0.0);
+
+            // Float inaccuracy does not guarantee approximate values
+            // In these tests, it diverges slightly over the epsilon value
+            //Test::assert_approx_eq(&pi.sin(), &0.0);
+            //Test::assert_approx_eq(&half.sin(), &1.0);
+
+            Test::assert_approx_eq(&zero.tan(), &0.0);
+            Test::assert_approx_eq(&nzero.tan(), &0.0);
+
+            //Test::assert_approx_eq(&pi.tan(), &0.0);
+            // Cannot test for PI / 2 equality as it diverges to infinity
+            // Float inaccuracy does not guarantee a NAN or INFINITY result
+            //let result = half.tan();
+            //assert!(result == V::nan() || result == V::infinity());
+
+            Test::assert_approx_eq(&zero.cosh(), &1.0);
+            Test::assert_approx_eq(&nzero.cosh(), &1.0);
+
+            Test::assert_approx_eq(&zero.sinh(), &0.0);
+            Test::assert_approx_eq(&nzero.sinh(), &0.0);
+
+            Test::assert_approx_eq(&zero.tanh(), &0.0);
+            Test::assert_approx_eq(&nzero.tanh(), &0.0);
+        }
+    }
+}
